max/min need to be Double now.
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@20492 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/query/param/NumberRangeParam.java
@@ -22,8 +22,8 @@ import org.json.JSONObject;
 public class NumberRangeParam extends Param {
 	
   private Integer numDecimalPlaces = new Integer(1);
-  private Long min;
-  private Long max;
+  private Double min;
+  private Double max;
   private boolean integer;
 
   private List<WdkModelText> regexes;
@@ -222,19 +222,19 @@ public class NumberRangeParam extends Param {
 	this.numDecimalPlaces = numDecimalPlaces;
   }
 
-  public Long getMin() {
+  public Double getMin() {
 	return min;
   }
 
-  public void setMin(Long min) {
+  public void setMin(Double min) {
 	this.min = min;
   }
 
-  public Long getMax() {
+  public Double getMax() {
 	return max;
   }
 
-  public void setMax(Long max) {
+  public void setMax(Double max) {
 	this.max = max;
   }
